requirements.txt
streamlit
pillow
easyocr
deep-translator
streamlit-drawable-canvas
opencv-python
